// web/src/models/apiResponse.model.ts
export interface apiResponseData {
  avatar_url: string;
  bio: string;
  blog: string;
  company: string | boolean;
  created_at: string;
  email: string | null;
  events_url: string;
  followers: number;
  followers_url: string;
  gists_url: string;
  gravatar_id: string;
  hireable: boolean;
  html_url: string;
  id: number;
  location: string;
  login: string;
  name: string;
  node_id: string;
  organizations_url: string;
  public_gists: number;
  public_repos: number;
  received_events_url: string;
  repos_url: string;
  site_admin: boolean;
  starred_url: string;
  subscriptions_url: string;
  twitter_username: string | null;
  type: string;
  updated_at: string;
  url: string;
}

export interface apiResponseRepositories {
  id: number;
  node_id: string;
  name: string;
  full_name: string;
  private: boolean;
  owner: {
    login: string;
    id: number;
    node_id: string;
    avatar_url: string;
    gravatar_id: string;
    url: string;
    html_url: string;
    followers_url: string;
    following_url: string;
    gists_url: string;
    starred_url: string;
    subscription_url: string;
    organizations_url: string;
    repos_url: string;
    events_url: string;
    received_url: string;
    type_url: string;
    site_admin: boolean;
  };
}